[dev-overlay] Move `error.name` to label
It's somewhat redundant to label errors with e.g.
"Runtime Error" and then also prefix the body with
"Error: ".

Now we just show the error name in the label, and
keep the body clean.

This should also teach better to look at the label
e.g. when you encounter `ReferenceError` you get
"Runtime ReferenceError" in the label.